Reenable CUDA version checks from Python.
These had been accidentally broken at some point in the plugin
switchover..
